Allows passing a factory for producing a stream to the ImplicitHeadMiddleware constructor
Instead of hard-coding generation of a stream, you may now pass a PHP
callable capable of producing a PSR-7 `StreamInterface` instance to the
`ImplicitHeadMiddleware` constructor, as the second argument.

If none is provided, it will create a closure that returns a
zend-diactoros `Stream` instance, backed by a writable `php://temp`
stream.